build: make postcss works in both snowpack and storybook
