+++ jest.config.js
@@ -59,6 +59,7 @@ module.exports = {
     '/packages/db/.shadow-cljs/',
     '/packages/spec/src/spec.js',
     '/packages/spec/.shadow-cljs/',
+    '/packages/storybook',
   ],
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
@@ -86,6 +87,7 @@ module.exports = {
     '/packages/db/',
     '/packages/bytes/',
     '/testlib/',
+    '/packages/storybook',
   ],
 
   // The directory where Jest should output its coverage files
