+++ packages/storybook/package.json
@@ -13,6 +13,7 @@
   "devDependencies": {
     "@babel/core": "^7.13.15",
     "@mdx-js/react": "^1.6.22",
+    "@storybook/addon-a11y": "^6.2.8",
     "@storybook/addon-docs": "^6.2.8",
     "@storybook/addon-essentials": "^6.2.8",
     "@storybook/addon-links": "^6.2.8",
